VBA Program thusfar
This is a simple program that I made for my mother and can  be updated by just changing out what is displayed in the Label, and from the IF statements.  Just images need updated.  I've been looking into changing the timer format, just hadn't had the chance to change it lately.  I'd love to change the .exe output to have a better program image, yet I haven't been able to figure out how to do this just yet.